requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
boto3==1.34.22
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==1.1.0
celery==5.3.6
redis==5.0.1
google-generativeai==0.5.4
PyPDF2==3.0.1
reportlab==4.4.5
pdf2image==1.17.0
pypdf==6.4.0
requests==2.32.3
Pillow==10.1.0


